Add advice for immediate termination with exit
exit() in Python mode has surprising behavior compared to the Java API, in that it continues to execute for the rest of setup or the rest of the draw loop. This is already documented, but the proposed addition adds a line of advice for people who want the immediate termination behavior typical in other modes.